v1.6

+++ services/notifications/notification_service.go
@@ -23,11 +23,11 @@ func AddNotification_Service(notification models.Mo_Notifications) (int, bool, s
 	return 201, false, "", "Notificacion agregado correctamente"
 }
 
-func ShowNotification_Service(iduser_int int, page_int int, idtype_int int) (int, bool, string, []*models.Mo_NotificationShow) {
+func ShowNotification_Service(iduser_int int, page_int int64, idtype_int int) (int, bool, string, []*models.Mo_NotificationShow) {
 
 	var notification_var []*models.Mo_NotificationShow
 
-	notifications, error_show := notification_repository.Mo_Find(iduser_int, int64(page_int), idtype_int)
+	notifications, error_show := notification_repository.Mo_Find(iduser_int, page_int, idtype_int)
 	if error_show != nil {
 		return 500, true, "Error en el servidor interno al intentar mostrar las notificaciones, detalles: " + error_show.Error(), notifications
 	}
